Extract the input validation and the `keyDown` handler from `index.js`
